Updates per 2/8 code review:  * do_chksum() returns size.  * connect.h contains all connection related headers.  * radstat gets finderinfo, type and applesingle size  * retr takes a size  * stor takes a size

+++ cksum.h
@@ -1,2 +1,2 @@
-int do_cksum( char *path, char *cksum_b64 );
-int do_acksum( char *path, char *cksum_b64, char *finfo_buf );
+ssize_t do_cksum( char *path, char *cksum_b64 );
+ssize_t do_acksum( char *path, char *cksum_b64, char *finfo_buf );
